Pad single digit numbers with zeros to a width of two
This change gives the time display a more standardised/common look.  When
using the app, I was surprised to see the seconds display go down to a
single digit and the display not to be padded by a zero, as this is not the
behaviour a standard digital clock has.  This is the reasoning behind this
commit: to reduce surprise to users.